[Marcion] Granular normalization flags.
Use granular flags for normalization, instead of a single universal flag
for all normalization types.

We, however, don't surface the flag all the way to the top-level code,
because the need for such granularity isn't currently evident. We're
introducing granularity just-in-case!

ⲛ̀ⲧⲁⲙϩϯ ⲛⲉⲙⲁⲕ, we should surface it using a parameter that looks like
this:
```
normalize: NormalizationConfig | bool
```
You can use `is_normalized_attestations=NormalizationConfig(...)` for
granularity, or `is_normalized_attestations=[True|False]` for
all-or-nothing.