[Marcion] "{nic}" shouldn't be appended!
"{nic}" is currently believed to mark spellings that are unattested in
Crum but presumed by Marcion. We remove it because those presumptions
are plausible (there are only two of them anyway!) And because its
current treatment as a word-wide detached type, rather than a
spelling-specific annotation, is going to be confusing.