[Kindle] Copy kindlegen binary from github.com/andyljones/kindlegen-64.
It is used in [1], which we intend to mimic. Amazon no longer supports
kindlegen, so we have to get the binary from elsewhere.

Amazon does publish a supposed alternative [2]. But I can't figure out
how to use it to generate a `.mobi` file. Running it against my EPUB
produces a KPF rather than a Mobi!

[1] https://github.com/tekinged/tekinged.com/blob/main/scripts/mk_kindle_dict.py
[2] https://www.amazon.com/Kindle-Previewer