[Crum] Run `make marcion_img`.
This regenerates many of the images with the previous change in
behavior. PNG background is manually set to white, thus retaining the
visual image after conversion to JPG.
(I am unsure why, but WEBP images have changed as well!)

Why are we changing PNG to JPG in the first place?
It's because we wan `run marcion_img` to be idempotent, meaning that
running it several times on the same input should produce identical
output. This is not the case for PNG images, which get modified (maybe
they store a timestamp there somewhere?) So we use JPG to make the
operaation idempotent.
If we can achieve idempotence with PNG, then conversion to PNG and this
(hacky) handling of the background color won't be necessary.