[Flashcards] Delete the Bible decks.
Supporting variable deck names significantly complicates our code.
And the Bible shouldn't be provided as a flashcards in the first place.
It likely won't be used at all. We didn't admit it back then, but we did
it because we simply felt the need to have everything as an Anki deck,
even if it didn't fit!