[Flashcards] Rename DIR format to TSV format.
We initially called it DIR because the metadata was written as a JSON
rather than a TSV, but DIR just doesn't make any sense! The bulk of the
data is in TSV. It's OK for the metadata to be written as a JSON and for
the format to continue to be called TSV. Just like our HTML output
directory will also contain CSS and media files, but we still call it
HTML because the core data will live in the HTML, and that's fine.