[Flashcards/KELLIA] Run `make flashcards test` to pick up changes.
The changes are supposed to affect formatting only. However, they seem
to have caused the flashcards pipeline to insert a lot of <br/> tags
unnecessarily.

Here is how this happened:
- We start wrapping the source XML files at 80 characters.
- Our KELLIA binary reads them, and passes the content as is to the
  output TSV's.
- Our Flashcards binary reads the TSV's, and replaces the newline
  characters with <br/> tags.

This looks like a bug that need to be resolved. Let's do it when we are
more fluent in KELLIA! (The code is mostly copied.)