[#184][Flashcards/Makefile] Rename the Test Rules.
In 42117938d5033e48ea318cff5f86c1b757a7b935, we started exporting an
Anki package in the test rules, but we retained the _html suffix, which
is misleading.

The only reason _html was introduced is that this is an auxiliary rule
(introduced simply because its presence is required due to the levels we
split our Makefile into), but we don't like the _aux suffix. We use an
underscore prefix instead.