[#82][Flashcards] Move Utils from field.py to constants.py.
That's where they belong.

Also modify the docs for the page deduplication method to avoid
confusion. Rename a parameter from `unique` to `at_most_once`, because
`unique` is used throughout the repo for something different - to force
entries to be unique. Thus use of this name for a different purpose is
confusing.