[#191][Site/Makefile] Restructure the Site Makefile Rules.
`site_prepublish` gives the impression that it runs before
`site_publish`. This is not the case, because the two recipes have an
overlap, and they're not meant to run in sequence.

`site_commit` is meant to create a commit that is available for manual
inspection prior to publishing.

`site_push` pushes the commit created in `site_commit`.

`site_publish` combines both.

We also add a `site_tidy` rule.
We don't include that in any of the other rules because it slows them
down unnecessarily. It is not desirable to run it every time,
because the HTML is usually pre-tidied, even with the changes made by
the publishing pipeline. So we add the site_tidy recipe in Makefile to
make it possible to invoke it manually for verification.

Also color the Output of an Error Message.