[Flashcards] Duplicate Crum and Dawoud Page Lists, and add Hyperlinks.
Let's say you've gone through the card, read the spellings, meaning,
looked at the explanatory images, and the derivations table, and you
ended up at the dictionary scans. Then you have to scroll all the way up
to find out where the words are exactly or on which page they're
defined. Having the page list right above the scans helps you quickly
find out where exactly the definitions are in the scans without having
to scroll all the way up.

We still retain the page list up next to the meaning in case you want to
scroll quickly to the scans while you're still up there. So no matter
where you are in the page, you can quickly go to the scans.

We also have hyperlinks in the first occurrence of the Crum and Dawoud
pages that lead to the occurrences at the bottom - not to any
particular page, but to the lists themselves.