[#188][Crum/HTML] Fix Bug. Width CSS Attribute is Always Present.
The previous code only set the attribute if the `id` parameter was set.
This didn't cause problems because `id` was always there. We add the
width attribute so it will be present even if `id` is set to false.

The bug was introduced in c4b05b77621562f2981d4ac21a37e50c28d9dbc2.

Running `make flashcards test` doesn't introduce any changes, as
expected, because this code path is currently unused as explained.