[#209][Crum/Flashcards] Crum Pages can be Absent. List all Pages Once.
The Crum page(s) occur(s) twice in the page. In the first time, we
restrict it to the location of the root, for consistency with other
dictionaries that reference Crum pages. In the second time, we list all
pages, to make lookup convenient, and for consistency with Dawoud. There
is no need to have consistency with Dawoud in the first occurrence.

Update the TypeScript to handle page numbers without the column. Run
`make ts_transpile` to regenerate the JavaScript.