[Crum] Delete Obsolete GSHEET Export Code.
We are starting to think that writing to a GSHEET isn't going to
happen. We use GSHEET files as a source rather than a sink. Although
it might be beneficial to have GSHEET sinks for visibility and
developer convenience. In such case, we will use a helper from
`utils`. We are therefore deleting the code from the Crum pipeline.

P.S. We have verified this commit by running `make crum test`. No
changes were introduced.