[#66][TypeScript] Set Version to ES2019.
`jshint` seems to support up to ECMAScript 11[1] (although it worked
fine when the version was set to `15`), which is supposedly equivalent
to `ES2020`. But we ran into some issues with `fixmyjs` when
we set the target to `ES2020`. (This hook has been problematic, and the
errors likely have to do with `fixmyjs` itself rather than anything that
we do.) So we downgraded the target to `ES2019`. It's good to have an
older target because then our generated code will be guaranteed to run
on more platforms.

The current setup is fine, so we use it as is.

[1] https://jshint.com/docs/options/#esversion