[#171,#229][Crum/Flashcards] Add a Line Break between Meaning and Crum.
This is obviously related to #171, but you might be wondering how it has
anything to do with #229.
The answer is that, with the current HTML, some text extractors see
"Crum" and whatever precedes it as a single word, thus messing up with
our search function (and also with text highlighting experience in the
browser). Adding a line break is a way to ensure that this doesn't
happen.