[#229][Crum/Search] Run `make xooxle`.
Pick up a change introduced in f7691b8f1ca20e664f7a85473d49d06179824b97,
and propagated to the HTML in f1114e6daf9df26510c487312d2168fbfb16ab0d
(we use HTML as a source for the `xooxle` recipe, so the code change
wouldn't be picked up otherwise). Examine with `diff` with the
`--word-diff` command to see the exact fix.