[#9][Crum/Typos] Fix Typo.
Anecdotally, the typo was discovered thanks to a pre-commit hook, which
normally doesn't run on text files, but we included the text for 1
(ⲟⲩⲱⲓⲛⲓ) as an example for local testing in our `xooxle.ts`, so the typo
was caught by the hook!
We might want to consider running the English subset of our data against
a spelling checker.