[#229][Crum/Search] Start Search at Checkbox Check Events as Well.
This slightly improves the experience, so users don't have to click the
button at all.

So far, we have the following triggers:
- Clicking the search button.
- Pressing enter while in the textbox.
- Checking any of the checkboxes.

Ideally, we would get rid of the button, and just work with any
`keypress` event on the textbox.
We can't do this right now because it seems to lag. Every time we add a
letter, search seams to work with the text taken before the new letter,
rather than with the new letter!