[#229][Crum/Search] Page Link Lives on a Separate Column.
Previously, we used to embed the link the word. This made it ugly due to
the link format that it acquires.
We considered hiding the link format, but we do this at the risk of new
users not discovering that the links exist and that they can retrieve
more context by clicking it.
We opted for storing the links to the pages on a third column.