[#229,#233][Crum/Flashcards] Use `<div>` Tags more Extensively.
We split the page into sections represented by `<div>` tags. This will
make the HTML prettier (tidy handle `<div>` tags intelligently). It will
also come in handy in the future when we want to extract specific
sections from the HTML, as we want to do in #229.

Notice that the use of `<div>` tags makes a lot of `<br>` tags
idempotent, so we remove them.