[#233,#171][Crum/Flashcards] Insert a `<br>` if Images are Absent.
This existed, but was mistakenly removed in
f9c7165db921164b6ea3e55f05f6befaa953dd9f, perhaps under the assumption
that the use of `<div>` tags would render the line breaks unnecessary.
It's necessary to have two line breaks, not just one, when the images
are absent; in order for the horizontal line to avoid conflict with the
"dictionary" division.