[#229][Crum/Search] Fix Bug. Add Event Listener to the Regex Checkbox.
The code had two listeners on the full-word checkbox, but none for the
use-regex checkbox!

The bug was introduced when the responsiveness was first implemented in
dc1d809949ea109b4f0878d501b2f15a2700e296, meaning that the regex
checkbox never had the event attached.
The value of the regex checkbox was still respected, it just didn't
trigger a display updated when changed.