[#5][Crum/Images] Fix 639-1-1.
We manually convert from PNG to JPG, and then provide a JPG as the
source image.
We can't modify our `magick` command to a one-size-fits-all for PNG
images. The command that worked for this particular image doesn't
work for others!

For the record, this is that worked for 639-1-1:
```
magick ${SRC} -background white -flatten -alpha off ${DST}
```