[#5][Crum/Images] Support Colors in `crum_img_plot`.
Use `less -R` instead of just `less`.
On your local machine, `less` is aliases to `less -R`, so it doesn't
matter. This isn't necessarily the case elsewhere, so we explicitly add
the flag in our invocation.

This is a followup to 16a179884c1f80c68d39609b51fcb94cc6f984e8.