[#9,#5][Crum/Xooxle] Run `make xooxle` to Rebuild the Index.
This picks up some typos, but mainly image captions!
The inclusion of captions in the search index is intentional. Although
it is usually the case that the caption text is taken from existing text
in Crum's dictionary, this isn't always the case, thus adding the
caption text gives us more text to search.