[fix #22][Crum/Images/Helper] Query Freepik, Vecteezy, and Flaticon.
1. Add new icon sources - Flaticon, Vecteezy, and Freepik.

2. Treat Google as just another source.
We almost end up never fetching images from Google, or we at least
modify the query, because the meaning column is often cluttered with
text that shouldn't belong to the query. It is therefore more
appropriate to open it only upon request, just like we do with the other
sources; and with a query provided by the user, instead of simply
copying the entire meaning column.