[#266][CSS] Pick up the New Shared CSS.
NOTE: While this should normally be done by `make flashcards test`, we
can't do it that easily right now because the flag values in
`flashcards/main.py` are intentionally messed up due to #267. We
manually copied the files, to the same effect.

Running `todo 267` should point you to the code TODO's associated with
the issue.