[#226,#269,#271,#272][Crum/Sisters] Fix Bug in Helper Script.
While updating the sheet, read updated data from the sheet rather than the
local TSV. The TSV can be outdated.
In particular, the TSV can be updated after the sheet is updated during
a prompt session!
Running the following will result in bugs:
```
-s 1 2
-s 1 3
```
This will result in the following:
```
KEY SISTERS
1   3
2   1
3   1
```
The first entry should be `1   2;3`, but that TSV would be unaware of
`2` because we didn't add it to the TSV, we added it to the sheet.

Side change:
- Enforce homonym symmetry as well.