[#241,#202][Crum/Flashcards] Fix Bug. Fix Dictionary Page Image Logic.
The class was repurposed in 93440c98e8b712a6497eecba138381f790cce7d7
such that, it no longer refers to the `<img>` tag, but rather the
enclosing `<figure>` tag! We therefore update the code to work on the
first child of the tag bearing the classes, rather than on the tag
itself.

The buggy commit fixed the logic for explanatory images, but missed
dictionary pages!

Side change:
- Stop using the deprecated `substr` method.
- Use the new `chopColumn` method for both Dawoud and Crum pages, in
  case we start using `Dawoud` pages without the column suffix.