[#278][Crum/TypeScript] Fix Bug. `dev` Button also Updates CSS.
The current logic for `dev` only updated the local storage variable, but
not the CSS. It might have also been broken, because it seemed to not
even have that effect (maybe `el.onclick = dev;` didn't work)! We know
it likely didn't update the local storage variable because we tried
clicking the `dev` button and then refreshing the page to prompt a
display refresh, but the display remained the same, suggesting that the
local storage variable was not updated.