[#226,#269,#271,#272][Crum/Sisters] Fix Validation Bug.
The bug was introduced in 992158d506ac2e1d5e1c390aa13145efacb62f01. The
signature of `family.all_except_you` and `family.natives_except_you` has
changed to return a list of `person` objects, rather than a list of
strings. This breaks the validation code, which expects a list of strings.