[#241][Crum/TypeScript] Use Anki Highlighting when Failing to read CSS.
Our logic used to fail with this error when run locally *without* a
server:
```
script.js:51 Uncaught SecurityError: Failed to read the 'cssRules'
property from 'CSSStyleSheet': Cannot access rules
    at new Highlighter (script.js:51:42)
    at main (script.js:387:25)
    at script.js:679:3
```
This is due to CORS. We tried disabling CORS in
2e328af91b28ee645671a039c9a7be94a12c7570, but this ended up causing
other problems, so we reverted that commit. Now we try to work around
the failure by resorting to the Anki highlighting when failing to read
CSS rules in the JavaScript.