[#229][Xooxle] Fix Highlighting Bug.
The previous code used to surround a matching substring with a `<span>`
tag, regardless of the content of the string. The problem is that the
string could contain any number of opening / closing HTML tags, which
could conflict with our highlighting tag!

We now surround the text matches with the highlighting `<span>` tags,
thus guaranteeing that any conflicts with existing HTML tags are
avoided.

Side change:
- Highlight using a new `match` class that has a corresponding CSS rule,
  not manually the Xooxle code.