[#294][Crum/TypeScript] Explicitly Specify the Page Class in the HTML.
Previously, our Crum logic used to run in just two places: XOOXLE, and
NOTE.
We therefore defined a global variable, XOOXLE, to distinguish the two.
We define a NOTE page as a page where the XOOXLE variable is absent.

We are starting to have more classes, so we need each class to be
explicitly given.