[#368,#356][GitHub/Antinoou] Create `.nojekyll` File in `docs` Folder.
Jekyll, by default, ignores files or folders that begin with an
underscore. We have to prepend an underscore to our font files during
the pipeline changes to support Antinoou on Anki. Since we don't use any
Jekyll features anyway, we might as well just disable it, which will
also speed up the build.